% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterPredict.R
\name{afterPredict}
\alias{afterPredict}
\title{afterPredict}
\usage{
afterPredict(
  omic_test,
  omic_test_diagnosis,
  seed,
  dir_path,
  file_suffix,
  model,
  model_suffix,
  save_results = TRUE
)
}
\arguments{
\item{omic_test}{Data frame or matrix containing the test dataset.}

\item{omic_test_diagnosis}{Independent test dataset variable.}

\item{seed}{Seed for reproduciblity.}

\item{dir_path}{Directory where results should be saved.}

\item{file_suffix}{String suffix for naming result files.}

\item{model}{Caret trained model.}

\item{model_suffix}{String prefix for naming result files.}

\item{save_results}{Logical. Should the confusion matrix and model be saved? Default is TRUE.}
}
\value{
A list containing the confusion matrix and predictions.
}
\description{
This function return predicts results of a caret trained model and a test dataset and
optionally saves the results and model to specified file paths.
}
