% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping_shuffle.R
\name{bootstrapping_shuffle}
\alias{bootstrapping_shuffle}
\title{Bootstrapping Shuffle Function}
\usage{
bootstrapping_shuffle(clinic_data, svm_data, seed, saving_name)
}
\arguments{
\item{clinic_data}{Clinical dataset.}

\item{svm_data}{Transcriptomic dataset containing misclassified samples and their distances.}

\item{seed}{Seed for reproducibility.}

\item{saving_name}{String representing the base name for saving output files.}
}
\value{
The function saves bootstrapping results, thresholds,
and selected samples in the specified output directory.
}
\description{
Perform a bootstrapping-based shuffle process.
The function performs the following steps:
- Normalizes the distances
- Repeats the process 100,000 times, assigning random labels between Control and Case
  and calculates the threshold of that distribution
- Computes the average of all thresholds
- Selects the anomalous samples based on the average of the thresholds
}
