% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering function}
\usage{
clustering(
  omic_data,
  class_variable,
  changed_clinic_data,
  id_column,
  saving_name,
  changed_class,
  class,
  mtry,
  n_cores,
  seed,
  enrichment
)
}
\arguments{
\item{omic_data}{Transcriptomic data frame.}

\item{class_variable}{Stringname representing the class in the data.}

\item{changed_clinic_data}{Clinic dataset with MLASDO changed class variable.}

\item{id_column}{String name representing the sample ids column name.}

\item{saving_name}{String name used for naming saved output files.}

\item{changed_class}{Name of the change to be specifically analyzed 'Control2Case' or 'Case2Control'.}

\item{class}{Secondary change class: 'Case' for 'Case2Control' or 'Control' for 'Control2Case'.}

\item{mtry}{Number of variables to split at each node in random forest.}

\item{n_cores}{Number of cores to use in random forest training.}

\item{seed}{Seed for reproducibility.}

\item{enrichment}{Logical flag indicating if enrichment analysis should be performed.}
}
\value{
Clustering, feature importance and enrichment RDS files in a specified directory.
}
\description{
This function performs hierarchical clustering, performs feature importance
based on random forest model training and enrichment analysis most importance features.
It saves all clustering, features importance and enrichmnent results and outputs for downstream analysis.
}
