% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{.mlasdo_run_rf_feature_selection}
\alias{.mlasdo_run_rf_feature_selection}
\title{Run the Random Forest feature selection step}
\usage{
.mlasdo_run_rf_feature_selection(
  data,
  n_cores,
  clinical_data = NULL,
  id_column,
  target_column,
  always_split_variables,
  rf_mtry,
  rf_splitrule,
  rf_min_node_size,
  train_method,
  train_number,
  train_repeats,
  rf_num_trees,
  rf_model_path,
  seed,
  importance,
  verbose
)
}
\description{
Run the Random Forest feature selection step
}
