% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeGA.R
\name{executeGA}
\alias{executeGA}
\title{executeGA}
\usage{
executeGA(
  mlAlgorithm,
  numLassoExecutions,
  numTrees,
  mtry,
  splitrule,
  sampleFraction,
  maxDepth,
  minNodeSize,
  omicData,
  activePredictors,
  classVariable,
  savingName,
  nCores,
  partitionPercentage,
  nIterations,
  nStopIter,
  populationSize,
  diagnosticChangeProbability,
  crossoverOperator,
  crossoverProbability,
  selectionOperator,
  mutationOperator,
  mutationProbability,
  seed
)
}
\arguments{
\item{mlAlgorithm}{Machine Learning algorithm to be applied, the options are: Lasso or RF (Random Forest).}

\item{numLassoExecutions}{Number of times the Lasso algorithm is executed.}

\item{numTrees}{Number of trees of the Random Forest model.}

\item{mtry}{Number of predictors that are evaluated at each partition (node) of each tree.}

\item{splitrule}{This is the rule used by the algorithm to select the predictor and the optimal value to separate a node into two branches during the tree construction.}

\item{sampleFraction}{Fraction of the training data that will be used to create each of the trees in the forest.}

\item{maxDepth}{Maximum height of each tree in the forest.}

\item{minNodeSize}{Minimum number of observations required in a node to be able to split it.}

\item{omicData}{Dataset of omic data that will be used.}

\item{activePredictors}{Predictors on which the study of the ratios will be conducted after the genetic algorithm has been performed.}

\item{classVariable}{Target variable, which must be binary, meaning it has two possible values.}

\item{savingName}{Name under which the model and solution will be saved after execution.}

\item{nCores}{Number of cores to be used in parallelization.}

\item{partitionPercentage}{Percentage (expressed as a fraction) with which the data will be split into a training and test set.}

\item{nIterations}{Number of iterations (generations) the genetic algorithm will perform.}

\item{nStopIter}{Number of iterations after which the algorithm will stop if all of them have the same fitness value.}

\item{populationSize}{Number of solutions that will be part of the initial population.}

\item{diagnosticChangeProbability}{Percentage (expressed as a fraction) indicating the probability of each gene in the solutions to be changed.}

\item{crossoverOperator}{Crossover operator used in the genetic algorithm.}

\item{crossoverProbability}{Percentage (expressed as a fraction) indicating the probability of crossover occurrence.}

\item{selectionOperator}{Selection operator used in the genetic algorithm.}

\item{mutationOperator}{Mutation operator used in the genetic algorithm.}

\item{mutationProbability}{Percentage (expressed as a fraction) indicating the probability of mutation occurrence.}

\item{seed}{Seed used for the creation of training and test sets.}
}
\description{
Methodology based on the combination of a genetic algorithm and a Machine Learning technique to mutate the final
diagnosis of patients, detecting anomalies in them.
}
\details{
These discrepancies may have their origin in the evolution of the subject itself, leading it from one group to
another in biological terms, or they may derive from human error in the labelling of the samples, among other causes.

This detection is done by combining a genetic algorithm and a machine learning algorithm, in this case a Lasso model.

A genetic algorithm is a technique inspired by biological evolution that simulates the process of natural selection
and evolution of species, the execution of a genetic algorithm starts with a randomly generated initial solution set.

     In this case, it is assumed that certain patients are misdiagnosed, so the initial solution set will be created by
     randomly changing 10% of the patient diagnoses in each of these solutions.

In each generation (iteration) of the genetic algorithm, each of the solutions from the set of possible solutions is
used as a training dataset for a Lasso model, which will classify the patients and whose performance will determine
the fitness value (score) of the solution.

     The score of each solution is defined by the average of the balanced accuracies obtained after training
     5 Lasso models with that solution as training dataset.

After this evaluation, a ranking of the solutions is created:

     The two best solutions are combined to generate a new solution that is added to the set of possible solutions
     to the problem.

     Simultaneously, a certain number of the worst solutions are eliminated from the generation, i.e. they are no
     longer part of the set of possible solutions to the problem.

This process is repeated for the specified number of generations, until a certain predefined threshold is reached in
the score of the solutions or until the value of the best score of the set of solutions is repeated for a certain
specified number of generations.

This procedure is used to determine which samples were anomalous. The genetic algorithm locates these anomalous
samples and the machine learning algorithm evidences this detection, validating that these samples belong to the
alternative group.
}
\examples{

MLASDO::executeGA(mlAlgorithm = mlAlgorithm, numLassoExecutions = numLassoExecutions, numTrees = numTrees, mtry = mtry, splitrule = splitrule, sampleFraction = sampleFraction, maxDepth = maxDepth, minNodeSize = minNodeSize, omicData = omicData, activePredictors = activePredictors, classVariable = classVariable, savingName = savingName, nCores = nCores, partitionPercentage = partitionPercentage, nIterations = nIterations, nStopIter = nStopIter, populationSize = populationSize, diagnosticChangeProbability = diagnosticChangeProbability, crossoverOperator = crossoverOperator, crossoverProbability = crossoverProbability, selectionOperator = selectionOperator, mutationOperator = mutationOperator, mutationProbability = mutationProbability, seed = seed)
}
