% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureImportance.R
\name{feature_importance}
\alias{feature_importance}
\title{Feature Importance Calculation}
\usage{
feature_importance(
  omic_data,
  model_dir,
  solution,
  mtry,
  filename_suffix,
  saving_name,
  class_variable,
  partition_percentage,
  id_column,
  n_cores,
  seed,
  bool,
  num_trees
)
}
\arguments{
\item{omic_data}{Omics dataframe.}

\item{model_dir}{String of the directory to save the model.}

\item{solution}{Vector indicating the solution for transformation of control2case or case2control.}

\item{mtry}{Number of variables randomly sampled at each split in Random Forest models.}

\item{filename_suffix}{String of the suffix for saved file names.}

\item{saving_name}{Prefix for saved file names.}

\item{class_variable}{String of the class variable.}

\item{partition_percentage}{Percentage of data to partition for training.}

\item{id_column}{String of the IDs variable.}

\item{n_cores}{Number of cores for parallel processing.}

\item{seed}{Seed for reproducibility.}

\item{bool}{String indicating transformation type ("Control2Case" or "Case2Control").}

\item{num_trees}{Number of trees for Random Forest models.}
}
\value{
Dataframe with sorted feature importance.
}
\description{
This function calculates the feature importance using Random Forest models.
of control2case or case2control changes.
}
