% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/performStadisticAnalysisUserVariableClassification.R
\name{perform_statistic_analysis_user_variable_classification}
\alias{perform_statistic_analysis_user_variable_classification}
\title{Perform Statistical Analysis for User Variable Classification}
\usage{
perform_statistic_analysis_user_variable_classification(
  saving_name,
  changed_clinic_data,
  first_group,
  second_group,
  active_predictors,
  categoric_active_predictors,
  numeric_active_predictors,
  class_variable,
  default_odds_predictors,
  method,
  dir_path
)
}
\arguments{
\item{saving_name}{Name to save the output files.}

\item{changed_clinic_data}{Modified clinic dataset for analysis.}

\item{first_group}{Name of the first group for classification.}

\item{second_group}{Name of the second group for classification.}

\item{active_predictors}{List of all predictors used in the analysis.}

\item{categoric_active_predictors}{List of categorical predictors.}

\item{numeric_active_predictors}{List of numeric predictors.}

\item{class_variable}{Name of the variable defining the classification groups.}

\item{default_odds_predictors}{Predictors for which default odds ratio calculations apply.}

\item{method}{Statistical method to apply (e.g., Fisher's test, Chi-square).}

\item{dir_path}{Path to the directory for saving results.}
}
\value{
Saves results including odds ratios and Wilcoxon test results to specified directory.
}
\description{
This function performs statistical analysis to evaluate the relationships between predictors and user-defined classes.
It calculates odds ratios, statistical tests, and Wilcoxon tests for categorical and numeric predictors.
}
