% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/performStadisticAnalysisAutomatedVariableClassification.R
\name{performStadisticAnalysisAutomatedVariableClassification}
\alias{performStadisticAnalysisAutomatedVariableClassification}
\title{performStadisticAnalysisUserVariableClassification}
\usage{
performStadisticAnalysisAutomatedVariableClassification(
  savingName,
  changedClinicData,
  firstGroup,
  secondGroup,
  activePredictors,
  categoricActivePredictors,
  numericActivePredictors,
  classVariable,
  defaultOddsPredictors
)
}
\arguments{
\item{savingName}{String | Name under which the model and solution will be saved after execution. If the user does not set any name, it will create a string with the current date.}

\item{changedClinicData}{Dataset | Dataset of clinic data that will be used.}

\item{firstGroup}{String | String that contains the name of the class created for the firstGroup. Case -> Control}

\item{secondGroup}{String | String that contains the name of the class created for the secondGroup. Control -> Case}

\item{activePredictors}{Array of Strings | Predictors on which the study of the ratios will be conducted after the genetic algorithm has been performed. Default value: All the predictors in clinic data, except classVariable and idColumn.}

\item{categoricActivePredictors}{Array of Strings | Categoric predictors of the active predictor list. Default value: All predictors in the active predictor list that return FALSE in the is.numeric function or return TRUE in the is.integer function and have less than 7 distinct values.}

\item{numericActivePredictors}{Array of Strings | Numerical predictors of the active predictor list. Default value: All predictors in the active predictor list that return TRUE in the is.numeric function or return TRUE in the is.integer function and have 7 or more distinct values.}

\item{classVariable}{String | Target variable, which must be binary, meaning it has two possible values. If the user does not specify a path to his own data, the value for the sample data, Ca.Co.Last, will be used.}

\item{defaultOddsPredictors}{String | Vector of names of categorical features that have a default value to compare this value against the rest of the values of that feature.}
}
\description{
This function performs ratio-based analysis on the active activePredictors passed as parameters.
}
\examples{

MLASDO::performStadisticAnalysisAutomatedVariableClassification(savingName = savingName, changedClinicData = changedClinicData, firstGroup = firstGroup, secondGroup = secondGroup, activePredictors = activePredictors, categoricActivePredictors = categoricActivePredictors, numericActivePredictors = numericActivePredictors, classVariable = classVariable, defaultOddsPredictors = defaultOddsPredictors)
}
