% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessChangedData.R
\name{preprocess_changed_data}
\alias{preprocess_changed_data}
\title{Preprocess Changed Data}
\usage{
preprocess_changed_data(
  omic,
  clinic_data,
  solution,
  active_predictors,
  class_variable,
  id_column,
  seed,
  svm_data
)
}
\arguments{
\item{omic}{Data frame containing omic data.}

\item{clinic_data}{Data frame containing clinical data.}

\item{solution}{Integer indicating the solution type (e.g., 1 for swapping classes).}

\item{active_predictors}{Vector of predictor variable names to exclude from the omic data.}

\item{class_variable}{String of the class variable.}

\item{id_column}{String of the IDs variable.}

\item{seed}{Seed for reproducibility.}

\item{svm_data}{List with SVM-related information, including `selected_samples$index`.}
}
\value{
A list containing four data frames:
  - `changed_omic_data`: Omic dataset modified according to `solution`.
  - `changed_clinic_data`: Clinical dataset modified according to `solution`.
  - `transition_clinic_data`: Clinical dataset with transition labels (e.g., "Case2Control").
  - `all_changed_omic_data`: Omic dataset where classes are changed only for the
    indices specified in `svm_data$selected_samples$index`.
}
\description{
This function preprocesses omic and clinical data based on the specified solution.
It also modifies the class variable and creates an additional dataset
`all_changed_omic_data` where class changes apply only to the indices
provided in `svm_data$selected_samples$index`.
}
