% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data for Training and Testing}
\usage{
preprocess_data(omic, class_variable, partition_percentage, id_column, seed)
}
\arguments{
\item{omic}{A data frame containing the omics dataset.}

\item{class_variable}{A string representing the name of the class variable column.}

\item{partition_percentage}{A numeric value specifying the proportion of the dataset to allocate to training (between 0 and 1).}

\item{id_column}{A string specifying the name of the ID column to be removed from the dataset.}

\item{seed}{An integer used to set the random seed for reproducibility.}
}
\value{
A list containing the following elements:
  \item{omic_train}{A data frame with training features.}
  \item{omic_train_diagnosis}{A vector of class labels for the training set.}
  \item{omic_test}{A data frame with testing features.}
  \item{omic_test_diagnosis}{A vector of class labels for the testing set.}
  \item{indices_train}{Indices of the rows used for training.}
}
\description{
This function partitions an omics dataset into training and testing subsets,
removing the class variable and a specified ID column from the feature set.
}
