% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainSVMRadial.R
\name{train_svm_radial}
\alias{train_svm_radial}
\title{Train an SVM Model with Radial Kernel}
\usage{
train_svm_radial(
  omic_train,
  omic_train_diagnosis,
  omic_test,
  omic_test_diagnosis,
  cost,
  seed,
  dir_path,
  file_suffix,
  n_cores
)
}
\arguments{
\item{omic_train}{Transcriptomic train dataframe with features (independent variables, x).}

\item{omic_train_diagnosis}{Vector of training data labels (dependent variables, y).}

\item{omic_test}{Transcriptomic test dataframe with features (independent variables, x).}

\item{omic_test_diagnosis}{Vector of test data labels (diagnoses).}

\item{cost}{Numeric value for the cost parameter of the SVM.}

\item{seed}{Seed for reproducibility.}

\item{dir_path}{Directory path where output files will be saved.}

\item{file_suffix}{Suffix to append to saved files.}

\item{n_cores}{Number of CPU cores to use for parallel processing.}
}
\value{
A list containing the trained SVM model, confusion matrix, balanced accuracy, and confidence intervals.
}
\description{
This function trains a Support Vector Machine (SVM) model with a radial kernel.
It evaluates the model using Balanced Accuracy (BA) and computes confidence intervals.
The trained model, confusion matrix, and confidence intervals are saved to specified files.
}
