% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileMarkdown.R
\name{compileMarkdown}
\alias{compileMarkdown}
\title{compileMarkdown}
\usage{
compileMarkdown(
  savingName,
  justAnalysis,
  lassoPredictorsPath,
  model,
  mlAlgorithm,
  geneticAlgorithm,
  originalDiagnosis,
  clinicData,
  classVariable
)
}
\arguments{
\item{savingName}{String | Name under which the model and solution will be saved after execution. If the user does not set any name, it will create a string with the current date.}

\item{justAnalysis}{Bool | Indicates whether to perform the analysis directly (TRUE) or to run the genetic algorithm (FALSE). Default value: FALSE.}

\item{lassoPredictorsPath}{String | Path to the mean number of predictors selected by Lasso in each generation.}

\item{model}{ML Model | Best model obtained after the detection.}

\item{geneticAlgorithm}{Array of Strings | Genetic algorithm object.}

\item{originalDiagnosis}{Array of Strings | Original diagnostics of the patients.}

\item{clinicData}{Data | Dataset of clinic data that will be used.}

\item{classVariable}{String | Target variable, which must be binary, meaning it has two possible values. If the user does not specify a path to his own data, the value for the sample data, Ca.Co.Last, will be used.}
}
\description{
This function compiles the markdown file with the analysis results
}
\examples{

MLASDO::compileMarkdown(savingName = savingName, mlAlgorithm = mlAlgorithm, lassoPredictorsPath = lassoPredictorsPath, model = model, geneticAlgorithm = geneticAlgorithm, clinicData = clinicData, classVariable = classVariable)

}
