% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileMarkdown.R
\name{compileMarkdown}
\alias{compileMarkdown}
\title{compileMarkdown}
\usage{
compileMarkdown(savingName, clinicData, classVariable)
}
\arguments{
\item{savingName}{Name under which the model and solution will be saved after execution. If the user does not set any name, it will create a string with the current date.}

\item{clinicData}{Dataset of clinic data that will be used.}

\item{classVariable}{Target variable, which must be binary, meaning it has two possible values.}
}
\description{
This function compiles the markdown file with the analysis results
}
\examples{

MLASDO::compileMarkdown(savingName = savingName, clinicData = clinicData, classVariable = classVariable)

}
