% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileMarkdown.R
\name{compileMarkdown}
\alias{compileMarkdown}
\title{compileMarkdown}
\usage{
compileMarkdown(
  savingName,
  justAnalysis,
  geneticAlgorithm,
  mlAlgorithm,
  numModelExecutions,
  lassoPredictorsPath,
  baselinePrecision,
  baselinePredictors,
  originalDiagnosis,
  clinicData,
  selectedData,
  classVariable,
  bestModelCM,
  worstModelCM
)
}
\arguments{
\item{savingName}{String | Name under which the model and solution will be saved after execution. If the user does not set any name, it will create a string with the current date.}

\item{justAnalysis}{Bool | Indicates whether to perform the analysis directly (TRUE) or to run the genetic algorithm (FALSE). Default value: FALSE.}

\item{geneticAlgorithm}{GA | Genetic algorithm object.}

\item{mlAlgorithm}{String | Machine Learning algorithm to be applied, the options are: Lasso or RF (Random Forest).}

\item{numModelExecutions}{Integer | Number of times the Lasso algorithm is executed. Default value: 5.}

\item{lassoPredictorsPath}{String | Path to the mean number of predictors selected by Lasso in each generation.}

\item{baselinePrecision}{Decimal | Baseline precision obtained with the model before the detection.}

\item{baselinePredictors}{Integer | Baseline predictors selected with the model before the detection.}

\item{originalDiagnosis}{Array of Strings | Original diagnostics of the patients.}

\item{clinicData}{Dataset | Dataset of clinic data that will be used.}

\item{selectedData}{Dataset | Dataset of omic data with only the predictors selected by the Lasso model.}

\item{classVariable}{String | Target variable, which must be binary, meaning it has two possible values. If the user does not specify a path to his own data, the value for the sample data, Ca.Co.Last, will be used.}

\item{bestModelCM}{Confusion Matrix | Confusion matrix of the best model obtained before the detection.}

\item{worstModelCM}{Confusion Matrix | Confusion matrix of the worst model obtained before the detection.}
}
\description{
This function compiles the markdown file with the analysis results
}
\examples{

MLASDO::compileMarkdown(savingName = savingName, justAnalysis = justAnalysis, geneticAlgorithm = geneticAlgorithm, mlAlgorithm = mlAlgorithm, numModelExecutions = numModelExecutions, lassoPredictorsPath = lassoPredictorsPath, baselinePrecision = baselinePrecision, baselinePredictors = baselinePredictors, originalDiagnosis = originalDiagnosis, clinicData = clinicData, selectedData = selectedData, classVariable = classVariable,  bestModelCM = bestModelCM, worstModelCM = worstModelCM)

}
