% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectAnomalies.R
\name{detectAnomalies}
\alias{detectAnomalies}
\title{executeAndPerformAnalysis}
\usage{
detectAnomalies(
  mlAlgorithm,
  numLassoExecutions = 5,
  numTrees = 100,
  mtry = 225,
  splitrule = "gini",
  sampleFraction = 1,
  maxDepth = 4,
  minNodeSize = 30,
  omicDataPath = "",
  clinicDataPath = "",
  idColumn,
  activePredictors = NULL,
  classVariable,
  savingName = "",
  nCores = 6,
  partitionPercentage = 0.9,
  nIterations = 200,
  nStopIter = 25,
  populationSize = 150,
  diagnosticChangeProbability = 0.1,
  crossoverOperator = gabin_spCrossover,
  crossoverProbability = 0.8,
  selectionOperator = gabin_tourSelection,
  mutationOperator = gabin_raMutation,
  mutationProbability = 0.1,
  seed = 1234
)
}
\arguments{
\item{mlAlgorithm}{Machine Learning algorithm to be applied, the options are: Lasso or RF (Random Forest).}

\item{numLassoExecutions}{Number of times the Lasso algorithm is executed.}

\item{numTrees}{Number of trees of the Random Forest model.}

\item{mtry}{Number of predictors that are evaluated at each partition (node) of each tree.}

\item{splitrule}{This is the rule used by the algorithm to select the predictor and the optimal value to separate a node into two branches during the tree construction.}

\item{sampleFraction}{Fraction of the training data that will be used to create each of the trees in the forest.}

\item{maxDepth}{Maximum height of each tree in the forest.}

\item{minNodeSize}{Minimum number of observations required in a node to be able to split it.}

\item{omicDataPath}{Path to omic data. If the user does not specify a path, the sample data will be used.}

\item{clinicDataPath}{Path to clinic data. If the user does not specify a path, the sample data will be used.}

\item{idColumn}{Variable that indicates the identifier of each patient in both datasets.}

\item{activePredictors}{Predictors on which the study of the ratios will be conducted after the genetic algorithm has been performed.}

\item{classVariable}{Target variable, which must be binary, meaning it has two possible values.}

\item{savingName}{Name under which the model and solution will be saved after execution. If the user does not set any name, it will create a string with the current date.}

\item{nCores}{Number of cores to be used in parallelization.}

\item{partitionPercentage}{Percentage (expressed as a fraction) with which the data will be split into a training and test set.}

\item{nIterations}{Number of iterations (generations) the genetic algorithm will perform.}

\item{nStopIter}{Number of iterations after which the algorithm will stop if all of them have the same fitness value.}

\item{populationSize}{Number of solutions that will be part of the initial population.}

\item{diagnosticChangeProbability}{Percentage (expressed as a fraction) indicating the probability of each gene in the solutions to be changed.}

\item{crossoverOperator}{Crossover operator used in the genetic algorithm.}

\item{crossoverProbability}{Percentage (expressed as a fraction) indicating the probability of crossover occurrence.}

\item{selectionOperator}{Selection operator used in the genetic algorithm.}

\item{mutationOperator}{Mutation operator used in the genetic algorithm.}

\item{mutationProbability}{Percentage (expressed as a fraction) indicating the probability of mutation occurrence.}

\item{seed}{Seed used for the creation of training and test sets.}
}
\description{
Methodology based on the combination of a genetic algorithm and a Machine Learning technique to mutate the final
diagnosis of patients, detecting anomalies in them.
}
\details{
These discrepancies may have their origin in the evolution of the subject itself, leading it from one group to
another in biological terms, or they may derive from human error in the labelling of the samples, among other causes.

The identification and correction of these erroneous situations becomes essential to preserve the integrity
and accuracy of the data.
}
\examples{

MLASDO::detectAnomalies(savingName = "DefaultExecutionLasso", mlAlgorithm = "Lasso", classVariable = "Ca.Co.Last", idColumn = "Trial")

MLASDO::detectAnomalies(savingName = "DefaultExecutionRF", mlAlgorithm = "RF", classVariable = "Ca.Co.Last", idColumn = "Trial")

MLASDO::detectAnomalies(savingName = "QuickExecution", mlAlgorithm = "Lasso", nIterations = 3, nStopIter = 1, populationSize = 20, classVariable = "Ca.Co.Last", idColumn = "Trial", activePredictors = c("sex", "age", "Mutation", "Ethnicity"))

MLASDO::detectAnomalies(savingName = "ExecutionWithOwnData", mlAlgorithm = "RF", omicDataPath = "./myOmicData.tsv", clinicDataPath = "./myClinicData.tsv", idColumn = "Patient.Id", nIterations = 3, populationSize = 10, classVariable = "Diagnosis", activePredictors = c("sex", "age", "Ethnicity"))
}
