% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectAnomalies.R
\name{detectAnomalies}
\alias{detectAnomalies}
\title{executeAndPerformAnalysis}
\usage{
detectAnomalies(
  justAnalysis = FALSE,
  solutionPath = "",
  geneticPath = "",
  bestModelPath = "",
  worstModelPath = "",
  lassoPredictorsPath = "",
  mlAlgorithm,
  predictorsToSelect = 15,
  numModelExecutions = 5,
  numTrees = 100,
  mtry = 225,
  splitrule = "gini",
  sampleFraction = 1,
  maxDepth = 4,
  minNodeSize = 30,
  omicDataPath = "",
  clinicDataPath = "",
  omicPredictorsToIgnore = NULL,
  clinicPredictorsToIgnore = NULL,
  idColumn,
  activePredictors = NULL,
  classVariable,
  savingName = "",
  nCores = 6,
  partitionPercentage = 0.9,
  nIterations = 200,
  nStopIter = 25,
  populationSize = 150,
  diagnosticChangeProbability = 0.1,
  crossoverOperator = gabin_spCrossover,
  crossoverProbability = 0.8,
  selectionOperator = gabin_tourSelection,
  mutationOperator = gabin_raMutation,
  mutationProbability = 0.1,
  seed = 1234
)
}
\arguments{
\item{justAnalysis}{Bool | Indicates whether to perform the analysis directly (TRUE) or to run the genetic algorithm (FALSE). Default value: FALSE.}

\item{solutionPath}{String | Path to genetic algorithm solution.}

\item{geneticPath}{String | Path to genetic algorithm object.}

\item{bestModelPath}{String | Path to the best model obtained.}

\item{worstModelPath}{String | Path to the worst model obtained.}

\item{lassoPredictorsPath}{String | Path to the mean number of predictors selected by Lasso in each generation.}

\item{mlAlgorithm}{String | Machine Learning algorithm to be applied, the options are: Lasso or RF (Random Forest).}

\item{predictorsToSelect}{Integer | Number of predictors to be selected from the most important predictors ranked by the RF model. This parameter is a integer number between 1 and the total number of predictors in the data. Default value: 15.}

\item{numModelExecutions}{Integer | Number of times the Lasso algorithm is executed. Default value: 5.}

\item{numTrees}{Integer | Number of trees of the Random Forest model. Default value: 100.}

\item{mtry}{Integer | Number of predictors that are evaluated at each partition (node) of each tree. Default value: 225.}

\item{splitrule}{String | This is the rule used by the algorithm to select the predictor and the optimal value to separate a node into two branches during the tree construction. Default value: gini.}

\item{sampleFraction}{Decimal | Fraction of the training data that will be used to create each of the trees in the forest. Default value: 1.}

\item{maxDepth}{Integer | Maximum height of each tree in the forest. Default value: 4.}

\item{minNodeSize}{Integer | Minimum number of observations required in a node to be able to split it. Default value: 30.}

\item{omicDataPath}{String | Path to omic data. If the user does not specify a path, the sample data will be used.}

\item{clinicDataPath}{String | Path to clinic data. If the user does not specify a path, the sample data will be used.}

\item{omicPredictorsToIgnore}{Array of Strings | Variables to be removed from the omic dataset. These will not be taken into account in the execution.}

\item{clinicPredictorsToIgnore}{Array of Strings | Variables to be removed from the clinic dataset. These will not be taken into account in the execution.}

\item{idColumn}{String | Variable that indicates the identifier of each patient in both datasets. If the user does not specify a path to his own data, the value for the sample data, Trial, will be used.}

\item{activePredictors}{Array of Strings | Predictors on which the study of the ratios will be conducted after the genetic algorithm has been performed. Default value: All the predictors in clinic data, except classVariable and idColumn.}

\item{classVariable}{String | Target variable, which must be binary, meaning it has two possible values. If the user does not specify a path to his own data, the value for the sample data, Ca.Co.Last, will be used.}

\item{savingName}{String | Name under which the model and solution will be saved after execution. If the user does not set any name, it will create a string with the current date.}

\item{nCores}{Integer | Number of cores to be used in parallelization. Default value: 6.}

\item{partitionPercentage}{Decimal | Percentage (expressed as a fraction) with which the data will be split into a training and test set. Default value: 0.9 (90%).}

\item{nIterations}{Integer | Number of iterations (generations) the genetic algorithm will perform. Default value: 200.}

\item{nStopIter}{Integer | Number of iterations after which the algorithm will stop if all of them have the same fitness value. Default value: 25.}

\item{populationSize}{Integer | Number of solutions that will be part of the initial population. Default value: 150.}

\item{diagnosticChangeProbability}{Decimal | Percentage (expressed as a fraction) indicating the probability of each gene in the solutions to be changed. Default value: 0.1 (10%).}

\item{crossoverOperator}{String | Crossover operator used in the genetic algorithm. Default value: Single Point Crossover.}

\item{crossoverProbability}{Decimal | Percentage (expressed as a fraction) indicating the probability of crossover occurrence. Default value: 0.8 (80%).}

\item{selectionOperator}{String | Selection operator used in the genetic algorithm. Default value: Tournament Selection.}

\item{mutationOperator}{String | Mutation operator used in the genetic algorithm. Default value: Random Mutation.}

\item{mutationProbability}{Decimal | Percentage (expressed as a fraction) indicating the probability of mutation occurrence. Default value: 0.1 (10%).}

\item{seed}{Integer | Seed used for the creation of training and test sets. Default value: 1234.}
}
\description{
Methodology based on the combination of a genetic algorithm and a Machine Learning technique to mutate the final
diagnosis of patients, detecting anomalies in them.

These discrepancies may have their origin in the evolution of the subject itself, leading it from one group to
another in biological terms, or they may derive from human error in the labelling of the samples, among other causes.

The identification and correction of these erroneous situations becomes essential to preserve the integrity
and accuracy of the data.
}
\examples{

MLASDO::detectAnomalies(savingName = "DefaultExecutionLasso", mlAlgorithm = "Lasso",)

MLASDO::detectAnomalies(savingName = "DefaultExecutionRF", mlAlgorithm = "RF")

MLASDO::detectAnomalies(savingName = "QuickExecution", mlAlgorithm = "Lasso", nIterations = 3, nStopIter = 2, populationSize = 20, activePredictors = c("sex", "age", "Mutation", "Ethnicity"))

MLASDO::detectAnomalies(savingName = "ExecutionWithOwnData", mlAlgorithm = "RF", predictorsToSelect = 0.3, omicDataPath = "./myOmicData.tsv", clinicDataPath = "./myClinicData.tsv", idColumn = "Patient.Id", nIterations = 3, populationSize = 10, classVariable = "Diagnosis", activePredictors = c("sex", "age", "Ethnicity"))

MLASDO::detectAnomalies(justAnalysis = TRUE, mlAlgorithm = "Lasso", geneticPath = "GA.rds", solutionPath = "GA_solution.rds", bestModelPath = "GA_Best_Model.rds", worstModelPath = "GA_Worst_Model.rds", lassoPredictorsPath = "GA_Lasso_Predictors.rds", savingName = "ExecutionWithOwnData", omicDataPath = "./myOmicData.tsv", clinicDataPath = "./myClinicData.tsv",idColumn = "Patient.Id",classVariable = "Diagnosis", activePredictors = c("sex", "age", "Ethnicity"))
}
