% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performPCAAnalysis.R
\name{performPCAAnalysis}
\alias{performPCAAnalysis}
\title{performPCAAnalysis}
\usage{
performPCAAnalysis(
  idColumn,
  omicData,
  activePredictors,
  classVariable,
  savingName
)
}
\arguments{
\item{idColumn}{Variable that indicates the identifier of each patient in both datasets.}

\item{omicData}{Dataset of omic data that will be used.}

\item{activePredictors}{Predictors on which the study of the ratios will be conducted after the genetic algorithm has been performed.}

\item{classVariable}{Target variable, which must be binary, meaning it has two possible values.}

\item{savingName}{Name under which the model and solution were saved after the execution of the genetic algorithm.}
}
\description{
This function performs PCA analysis on the changed diagnoses after the execution of the genetic algorithm.
}
\examples{

MLASDO::performPCAAnalysis(idColumn = idColumn, omicData = omicData, activePredictors = activePredictors, classVariable = classVariable, savingName = savingName)
}
