% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performPCAAnalysis.R
\name{performPCAAnalysis}
\alias{performPCAAnalysis}
\title{performPCAAnalysis}
\usage{
performPCAAnalysis(
  justAnalysis,
  solutionPath,
  idColumn,
  omicData,
  activePredictors,
  classVariable,
  savingName
)
}
\arguments{
\item{justAnalysis}{Bool | Indicates whether to perform the analysis directly (TRUE) or to run the genetic algorithm (FALSE). Default value: FALSE.}

\item{solutionPath}{String | Path to genetic algorithm solution.}

\item{idColumn}{String | Variable that indicates the identifier of each patient in both datasets. If the user does not specify a path to his own data, the value for the sample data, Trial, will be used.}

\item{omicData}{Dataset of omic data that will be used.}

\item{activePredictors}{Array of Strings | Predictors on which the study of the ratios will be conducted after the genetic algorithm has been performed. Default value: All the predictors in clinic data, except classVariable and idColumn.}

\item{classVariable}{String | Target variable, which must be binary, meaning it has two possible values. If the user does not specify a path to his own data, the value for the sample data, Ca.Co.Last, will be used.}

\item{savingName}{String | Name under which the model and solution will be saved after execution. If the user does not set any name, it will create a string with the current date.}
}
\description{
This function performs PCA analysis on the changed diagnoses after the execution of the genetic algorithm.
}
\examples{

MLASDO::performPCAAnalysis(justAnalysis = justAnalysis, solutionPath = solutionPath, idColumn = idColumn, omicData = omicData, activePredictors = activePredictors, classVariable = classVariable, savingName = savingName)
}
