% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performRatioAnalysis.R
\name{performRatioAnalysis}
\alias{performRatioAnalysis}
\title{performRatioAnalysis}
\usage{
performRatioAnalysis(clinicData, activePredictors, classVariable, savingName)
}
\arguments{
\item{clinicData}{Dataset of clinic data that will be used.}

\item{activePredictors}{activePredictors on which the study of the ratios will be conducted after the genetic algorithm has been performed.}

\item{classVariable}{Target variable, which must be binary, meaning it has two possible values.}

\item{savingName}{Name under which the model and solution were saved after the execution of the genetic algorithm.}
}
\description{
This function performs ratio-based analysis on the active activePredictors passed as parameters.
}
\examples{

MLASDO::performRatioAnalysis(clinicData = clinicData, activePredictors = activePredictors, classVariable = classVariable, savingName = savingName)
}
