% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performRatioAnalysis.R
\name{performRatioAnalysis}
\alias{performRatioAnalysis}
\title{performRatioAnalysis}
\usage{
performRatioAnalysis(
  savingName,
  changedClinicData,
  firstGroup,
  secondGroup,
  activePredictors,
  classVariable
)
}
\arguments{
\item{savingName}{String | Name under which the model and solution will be saved after execution. If the user does not set any name, it will create a string with the current date.}

\item{changedClinicData}{Dataset | Dataset of clinic data that will be used.}

\item{firstGroup}{String | String that contains the name of the class created for the firstGroup. Case -> Control}

\item{secondGroup}{String | String that contains the name of the class created for the secondGroup. Control -> Case}

\item{activePredictors}{Array of Strings | Predictors on which the study of the ratios will be conducted after the genetic algorithm has been performed. Default value: All the predictors in clinic data, except classVariable and idColumn.}

\item{classVariable}{String | Target variable, which must be binary, meaning it has two possible values. If the user does not specify a path to his own data, the value for the sample data, Ca.Co.Last, will be used.}
}
\description{
This function performs ratio-based analysis on the active activePredictors passed as parameters.
}
\examples{

MLASDO::performRatioAnalysis(savingName = savingName, changedClinicData = changedClinicData, firstGroup = firstGroup, secondGroup = secondGroup, activePredictors = activePredictors, classVariable = classVariable)
}
